// src/constants/enums/accountSettings.ts
export enum AccountSettingForm {
  id = "account__setting_from",

  autocomplete_inputs = "off",
  input_size = "medium",
  button_size = "large",

  firstName_input_id = "firstName_input_id",
  firstName_input_name = "firstName_input",
  firstName_input_label = "",
  firstName_input_role = "firstName_input",
  firstName_input_arial_label = "first-name-input",
  firstName_input_placeholder = "Your first name",
  firstName_input_type = "text",

  lastName_input_id = "lastName_input_id",
  lastName_input_name = "lastName_input",
  lastName_input_label = "",
  lastName_input_role = "lastName_input",
  lastName_input_arial_label = "last-name-input",
  lastName_input_placeholder = "Your last name",
  lastName_input_type = "text",

  email_input_id = "email_input_id",
  email_input_name = "email_input",
  email_input_label = "",
  email_input_role = "email-input",
  email_input_arial_label = "email-input",
  email_input_placeholder = "Your email address",
  email_input_type = "email",

  password_input_id = "password_input_id",
  password_input_name = "password_input",
  password_input_label = "",
  password_input_role = "password-input",
  password_input_arial_label = "password-input",
  password_input_placeholder = "Your password",
  password_input_type_hidden = "password",
  password_input_type_show = "text",

  username_input_id = "username_input_id",
  username_input_name = "username_input",
  username_input_label = "",
  username_input_role = "username-input",
  username_input_arial_label = "username-input",
  username_input_placeholder = "Your username",
  username_input_type = "text",

  description_input_id = "description_input_id",
  description_input_name = "description_input",
  description_input_label = "",
  description_input_role = "description-input",
  description_input_arial_label = "description-input",
  description_input_placeholder = "Your description",
  description_input_type = "text",

  website_input_id = "website_input_id",
  website_input_name = "website_input",
  website_input_label = "",
  website_input_role = "website-input",
  website_input_arial_label = "website-input",
  website_input_placeholder = "Your website",
  website_input_type = "text",
}